fix(ci): line-limit guard — split scheduler helpers and worker tests
